requirements.txt
anykeystore==0.2
apex==0.9.10.dev0
appdirs==1.4.3
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptacular==1.5.5
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
efficientnet-pytorch==0.7.0
entrypoints==0.3
fastBPE==0.1.0
filelock==3.0.12
future==0.18.2
gudhi==3.3.0
html5lib==1.0.1
hupper==1.10.2
idna==2.8
imageio==2.9.0
ipaddr==2.2.0
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyterlab==2.2.0
jupyterlab-server==1.2.0
kiwisolver==1.2.0
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
msgpack==0.6.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.3.0.36
packaging==20.3
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
PasteDeploy==2.1.0
pbkdf2==1.3
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
piq==0.5.1
plaster==1.0
plaster-pastedeploy==0.7
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.6
pyramid==1.10.4
pyramid-mailer==0.15.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytorch-model-summary==0.1.2
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.1
regex==2020.6.8
repoze.sendmail==4.4.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.3.3
seaborn==0.10.1
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.14.0
SQLAlchemy==1.3.18
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.7.17
tokenizers==0.8.1rc1
torch==1.5.1
torchvision==0.6.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
transaction==3.0.0
transformers==3.0.2
translationstring==1.4
urllib3==1.25.8
velruse==1.1.1
venusian==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.6
widgetsnbextension==3.5.1
WTForms==2.3.1
wtforms-recaptcha==0.3.2
xlrd==1.2.0
zope.deprecation==4.4.0
zope.interface==5.1.0
zope.sqlalchemy==1.3
